hackathon_ET/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
requests==2.31.0
python-dotenv==1.0.1